Fix the rendered colors in the generated result web pages so that the colors are noticable for all non-noise results (0 - 2% is considered noise). The color maps were taken from the colorbrewer2.org web site, where we generated the color maps by:
  * Data classes:   9
  * Nature of Data: Sequential
  * Color Schemes:  Reds & Greens (from single hue selections)
  * Optionally:     showing only schemes that are colorblind safe
  * HEX values displayed for the color system

We then created an array of those values, throwing out the first given color
since it is so close to white, using white for a noisy positive or negative
value. Instead of using the factor to select the color, we set the color using
the percentage increase or decrease to select among 8 buckets. The buckets are
as follows (positive or negative uses the same index, just a different color
map):

  * Bucket Idx 1:   0.0% <= x <   2.0% (noise, white)
  * Bucket Idx 2:   2.0% <= x <  16.0%
  * Bucket Idx 3:  16.0% <= x <  30.0%
  * Bucket Idx 4:  30.0% <= x <  44.0%
  * Bucket Idx 5:  44.0% <= x <  58.0%
  * Bucket Idx 6:  58.0% <= x <  72.0%
  * Bucket Idx 7:  72.0% <= x <  86.0%
  * Bucket Idx 8:  86.0% <= x < 100.0%
  * Bucket Idx 9: 100.0% <= x

Note that any increase or decrease 100% or above is lumped into one color
since we want to highlight differences under 100%. It was felt that using more
colors than suggested by the cartographers was not wise, as little is gained
to have more than 8 colors.