Fix a bug where the exit.c benchmark was leaving 100s of processes around. This was due to the fact that the frame work could call the benchmark_initbatch() routine, but not call the benchmark() routine (since the time limit had been reached). For exit.c, this meant that all the processes as a result of the call to its benchmark_initbatch() were left around, since only in benchmark() are they reaped.
The fix was to change the frame work to revert to the previous behavior
(pre-51d7a006a6220e3f9caf6ecdf468b69591546f3c), where we would only check for
time or operation counts after a benchmark was taken.

Additionally, because both exit.c and fork.c use the barrier constructs from
the frame work, we ensured that the internal frame work data is only
considered when the frame work reports results to the barrier and the frame
work has setup those time checks. We also changed barrier_destroy() to check
for errors and report them, and unmap the size for the barrier that was
originally mapped.