Add summary, break out percent, show/hide columns
Some times the data being compared is from successive runs of the same
configuration. In that case, we don't need a comparison against the first set,
but a summary of the runs. This commit adds a summary set of data (min, max,
median, average, standard deviation) of the data sets being compared, and the
ability to show/hide that data.

We also breakout the percent increase/decrease column to be more space
efficient for large numbers of columns, and add the ability to hide
the results themselves when we just want to see the summary data.

Signed-off-by: Peter Portante <peter.portante@redhat.com>